requirements.txt
flask==2.3.3
torch==2.0.1
torchvision==0.15.2
Pillow==10.0.1
numpy==1.24.3
opencv-python==4.8.1.78
python-dotenv==1.0.0
Werkzeug==2.3.7